Port project from java 8, netbeans 8.2 to a Modular Java 11 project using NetBeans 11.0, jdk-11.0.4 & javafx-sdk-11.0.2 using branch java11 until ready to be merged into master.
