+++ nbproject/private/private.properties
@@ -1,10 +1,8 @@
-auxiliary.org-netbeans-modules-projectapi.issue214819_5f_fx_5f_enabled=true
 compile.on.save=true
 do.depend=false
 do.jar=true
+do.jlink=true
 javac.debug=true
 javadoc.preview=true
-javafx.run.as=standalone
-javafx.run.inbrowser=<Default System Browser>
-javafx.run.inbrowser.path=C:\\Program Files\\Internet Explorer\\IEXPLORE.EXE
-user.properties.file=C:\\Users\\goodc\\AppData\\Roaming\\NetBeans\\8.2\\build.properties
+jlink.strip=false
+user.properties.file=C:\\Users\\goodc\\AppData\\Roaming\\NetBeans\\11.0\\build.properties
