+++ nbproject/private/private.xml
@@ -1,23 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <project-private xmlns="http://www.netbeans.org/ns/project-private/1">
-    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="3">
-        <file>
-            <url>src/backupgnucash/BackupGnuCashController.java</url>
-            <bookmark id="2">
-                <name/>
-                <line>717</line>
-                <key/>
-            </bookmark>
-            <bookmark id="3">
-                <name/>
-                <line>726</line>
-                <key/>
-            </bookmark>
-        </file>
-    </editor-bookmarks>
+    <editor-bookmarks xmlns="http://www.netbeans.org/ns/editor-bookmarks/2" lastBookmarkId="0"/>
     <open-files xmlns="http://www.netbeans.org/ns/projectui-open-files/2">
-        <group>
-            <file>file:/E:/Data/Java/NetBeansProjects/BackupGnuCash/src/backupgnucash/BackupGnuCashController.java</file>
-        </group>
+        <group/>
     </open-files>
 </project-private>
