+++ nbproject/project.properties
@@ -3,53 +3,64 @@ annotation.processing.enabled.in.editor=false
 annotation.processing.processors.list=
 annotation.processing.run.all.processors=true
 annotation.processing.source.output=${build.generated.sources.dir}/ap-source-output
-application.title=BackupGnuCash
-application.vendor=cgood
+application.title=BackupGnuCashJ11
+application.vendor=goodc
 build.classes.dir=${build.dir}/classes
 build.classes.excludes=**/*.java,**/*.form
 # This directory is removed when the project is cleaned:
 build.dir=build
 build.generated.dir=${build.dir}/generated
 build.generated.sources.dir=${build.dir}/generated-sources
+build.modules.dir=${build.dir}/modules
 # Only compile against the classpath explicitly listed here:
 build.sysclasspath=ignore
 build.test.classes.dir=${build.dir}/test/classes
+build.test.modules.dir=${build.dir}/test/modules
 build.test.results.dir=${build.dir}/test/results
-compile.on.save=true
-compile.on.save.unsupported.javafx=true
 # Uncomment to specify the preferred debugger connection transport:
 #debug.transport=dt_socket
 debug.classpath=\
     ${run.classpath}
+debug.modulepath=\
+    ${run.modulepath}
 debug.test.classpath=\
     ${run.test.classpath}
+debug.test.modulepath=\
+    ${run.test.modulepath}
+# Files in build.classes.dir which should be excluded from distribution jar
+dist.archive.excludes=
 # This directory is removed when the project is cleaned:
 dist.dir=dist
-dist.jar=${dist.dir}/BackupGnuCash.jar
 dist.javadoc.dir=${dist.dir}/javadoc
+dist.jlink.dir=${dist.dir}/jlink
+dist.jlink.output=${dist.jlink.dir}/BackupGnuCashJ11
 endorsed.classpath=
 excludes=
 includes=**
-# Non-JavaFX jar file creation is deactivated in JavaFX 2.0+ projects
-jar.archive.disabled=true
 jar.compress=false
-javac.classpath=\
-    ${javafx.classpath.extension}
+javac.classpath=
 # Space-separated list of extra javac options
 javac.compilerargs=
 javac.deprecation=false
+javac.external.vm=false
+javac.modulepath=\
+    ${libs.JavaFX11.classpath}
+javac.processormodulepath=
 javac.processorpath=\
     ${javac.classpath}
-javac.source=1.8
-javac.target=1.8
+javac.source=11
+javac.target=11
 javac.test.classpath=\
-    ${javac.classpath}:\
-    ${build.classes.dir}
+    ${javac.classpath}
+javac.test.modulepath=\
+    ${javac.modulepath}:\
+    ${build.modules.dir}
 javac.test.processorpath=\
     ${javac.test.classpath}
 javadoc.additionalparam=
 javadoc.author=false
 javadoc.encoding=${source.encoding}
+javadoc.html5=false
 javadoc.noindex=false
 javadoc.nonavbar=false
 javadoc.notree=false
@@ -58,66 +69,30 @@ javadoc.splitindex=true
 javadoc.use=true
 javadoc.version=false
 javadoc.windowtitle=
-javafx.application.implementation.version=1.0
-javafx.binarycss=false
-javafx.classpath.extension=\
-    ${java.home}/lib/javaws.jar:\
-    ${java.home}/lib/deploy.jar:\
-    ${java.home}/lib/plugin.jar
-javafx.deploy.adddesktopshortcut=false
-javafx.deploy.addstartmenushortcut=false
-javafx.deploy.allowoffline=true
-# If true, application update mode is set to 'background', if false, update mode is set to 'eager'
-javafx.deploy.backgroundupdate=false
-javafx.deploy.disable.proxy=false
-javafx.deploy.embedJNLP=true
-javafx.deploy.includeDT=true
-javafx.deploy.installpermanently=false
-javafx.deploy.permissionselevated=true
-# Set true to prevent creation of temporary copy of deployment artifacts before each run (disables concurrent runs)
-javafx.disable.concurrent.runs=false
-# Set true to enable multiple concurrent runs of the same WebStart or Run-in-Browser project
-javafx.enable.concurrent.external.runs=false
-# This is a JavaFX project
-javafx.enabled=true
-javafx.fallback.class=com.javafx.main.NoJavaFXFallback
-# Main class for JavaFX
-javafx.main.class=backupgnucash.BackupGnuCash
-javafx.preloader.class=
-# This project does not use Preloader
-javafx.preloader.enabled=false
-javafx.preloader.jar.filename=
-javafx.preloader.jar.path=
-javafx.preloader.project.path=
-javafx.preloader.type=none
-# Set true for GlassFish only. Rebases manifest classpaths of JARs in lib dir. Not usable with signed JARs.
-javafx.rebase.libs=false
-javafx.run.height=600
-javafx.run.width=800
-javafx.signing.blob=false
-javafx.signing.enabled=true
-javafx.signing.type=self
-# Pre-JavaFX 2.0 WebStart is deactivated in JavaFX 2.0+ projects
-jnlp.enabled=false
-# Main class for Java launcher
-main.class=backupgnucash.BackupGnuCash
-# For improved security specify narrower Codebase manifest attribute to prevent RIAs from being repurposed
-manifest.custom.codebase=*
-# Specify Permissions manifest attribute to override default (choices: sandbox, all-permissions)
-manifest.custom.permissions=
-manifest.file=manifest.mf
-meta.inf.dir=${src.dir}/META-INF
-mkdist.disabled=false
-native.bundling.enabled=false
+# The jlink additional root modules to resolve
+jlink.additionalmodules=
+# The jlink additional command line parameters
+jlink.additionalparam=
+jlink.launcher=true
+jlink.launcher.name=BackupGnuCash
+main.class=org.openjfx.BackupGnuCash
 platform.active=default_platform
-project.license=gpl30
-run.classpath=\
-    ${dist.jar}:\
-    ${javac.classpath}:\
-    ${build.classes.dir}
+run.classpath=
+# Space-separated list of JVM arguments used when running the project.
+# You may also define separate properties like run-sys-prop.name=value instead of -Dname=value.
+# To set system properties for unit tests define test-sys-prop.name=value:
+run.jvmargs=
+run.modulepath=\
+    ${libs.JavaFXMODS11.classpath}:\
+    ${javac.modulepath}:\
+    ${build.modules.dir}
 run.test.classpath=\
-    ${javac.test.classpath}:\
-    ${build.test.classes.dir}
+    ${javac.test.classpath}
+run.test.modulepath=\
+    ${javac.test.modulepath}:\
+    ${build.test.modules.dir}
 source.encoding=UTF-8
 src.dir=src
-test.src.dir=test
+src.dir.path=classes
+test.src.dir=src
+test.src.dir.path=tests
