+++ src/backupgnucash/classes/org/openjfx/ChangeLog.txt
@@ -1,0 +1,51 @@
+BackupGnuCash
+
+V# 2.0.0 28 Jul 2019 Chris Good
+
+    * Convert to Modular Java 11 project.
+
+V# 1.3.1 26 Feb 2019 Chris Good
+
+    * Linux Java 8 Help button causes hang: wrap the call to any AWT APIs
+      (Desktop.getDesktop().browse) in a runnable and submit it for execution
+      via java.awt.EventQueue.invokeLater().
+
+V# 1.3.0 24 Feb 2019 Chris Good
+
+    * Mods for GnuCash V3.
+    * Optionally back up V2 + V3 configuration folders and files.
+    * Backup more V2 configuration files, including Windows registry,
+      Linux dconf, and AqBanking settings.
+    * Add Help button.
+
+V# 1.22 10 Mar 2018 Chris Good
+
+    * Fix scr/backupgnucash/README.md multiword Table Of Contents entries
+    * Remove warning when running for first time and settings have not yet been
+      saved:
+       getUserDefaults: /home/[user]/.BupGc/defaultProperties (No such file or directory)
+
+V# 1.21 21 Jul 2016 Chris Good
+
+    * Fix use of & instead of && (didn't actually matter).
+    * Use netbeans 8.1.
+
+V# 1.20 15 Jul 2016 Chris Good
+
+    * Check for Windows 7-Zip in Program Files (x86).
+    * Allow configurations for multiple books.
+    * Make ToolTips font bigger.
+
+    Note: Java 1.8.0_72 or later is now required due to bug 
+        https://bugs.openjdk.java.net/browse/JDK-8136838
+        which meant the value of ComboBox.getValue() was not correct in previous
+        versions.
+        As of 15 Jul 2016, the current Java version on Windows is 1.8.0_92 and
+        on Ubuntu 16.04 is 1.8.0_91.
+        Ubuntu 16.04 openjfx is version 8u60-b27-4 which works so long as when 
+        adding a new book, ENTER is pressed after typing a new book name into
+        the Book combobox. I.e. Press ENTER before leaving the combobox.
+
+V# 1.10 28 May 2016 Chris Good
+
+    * Initial release
